requirements.txt
# Day 1 - Basic requirements
python-dotenv==1.0.0

# We'll add more as we go